works
I'm really surprised this worked the first time since I wasn't sure how insertion was supposed to work. I just tried to push to left or right based on size of the int (wouldn't the sum of child integers be more appropriate?), and returned the push function with the child node if it already exists.